make api zennotes

+++ app/Http/Controllers/Api/V1/NoteController.php
@@ -2,16 +2,16 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use Illuminate\Http\Request;
+use App\Models\Tag;
 use App\Models\Note;
 use App\Models\TagNote;
+use Illuminate\Http\Request;
+use App\Services\V1\NoteQuery;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\V1\NoteResource;
 use App\Http\Resources\V1\NoteCollection;
 use App\Http\Requests\V1\StoreNoteRequest; 
 use App\Http\Requests\V1\UpdateNoteRequest;
-use App\Services\V1\NoteQuery ;
-
 
 class NoteController extends Controller
 {
@@ -29,7 +29,18 @@ class NoteController extends Controller
      */
     public function index(Request $request)
     {
-        return $notes = $this->noteQuery->filterByTags($request);
+        $user = $request->user();
+
+        if (!$user) 
+        {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
+
+        $notes = $this->noteQuery->filterByTags($request) // get filter request: 
+            ->where('user_id', $user->id)   // localhost:8000/api/v1/notes?tags[]=tag1&tags[]=tag2
+            ->get();
+
+        return new NoteCollection($notes);
     }
 
     /**
@@ -40,6 +51,13 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id !== auth()->id()) 
+        {
+            return response()->json([
+                'message' => 'Not authorized.'
+            ], 403);
+        }
+    
         return new NoteResource($note);
     }
 
@@ -51,7 +69,33 @@ class NoteController extends Controller
      */
     public function store(StoreNoteRequest $request)
     {
-        return new NoteResource(Note::create($request->all()));
+        $note = Note::create([
+            'header' => $request->header,
+            'text_note' => $request->text_note,
+            'user_id' => $request->user()->id,
+        ]);
+
+        if ($request->has('tags')) 
+        {
+            foreach ($request->tags as $tag)
+            {   
+                if (Tag::where('tag_name', "=", $tag['tag_name'])
+                    ->where('user_id', "=", $request->user()->id)
+                    ->get()
+                    ->isEmpty())
+                {
+                    Tag::create([
+                        'tag_name' => $tag['tag_name'],
+                        'user_id' => $request->user()->id,
+                    ]);
+                }
+            }
+        }
+
+        $tags = Tag::whereIn('tag_name', $request->tags)->get();
+        $note->tags()->sync($tags);
+    
+        return new NoteResource($note);
     }
 
     /**
@@ -63,7 +107,15 @@ class NoteController extends Controller
      */
     public function update(UpdateNoteRequest $request, Note $note)
     {   
+        if ($note->user_id !== auth()->id()) 
+        {
+            return response()->json([
+                'message' => 'Not authorized.'
+            ], 403);
+        }
+    
         $note->update($request->all());
+        return new NoteResource($note);
     }
 
     /**
@@ -74,7 +126,25 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
-        TagNote::where('note_id', $note->id)->delete();
-        return Note::destroy($note->id);
+        try 
+        {
+            if ($note->user_id !== auth()->id()) 
+            {
+                return response()->json(['message' => 'Not authorized.'], 403);
+            }
+            
+            TagNote::where('note_id', $note->id)->delete();
+            Note::destroy($note->id);
+            return response()->json([
+                'message' => 'Note is deleted'
+            ]);
+        }
+        catch(\Throwable $th)
+        {
+            return response()->json([
+                'status' => false,
+                'message' => $th->getMessage()
+            ]);
+        }
     }
 }
